requirements.txt
cobs~=1.2.2
pyserial~=3.5
